+++ src/utility/circular_buffer.h
@@ -29,7 +29,7 @@ public:
             nb_read = (N > N_remain) ? N_remain : N;
         }
 
-        for (size_t i = 0; i < nb_read; i++) {
+        for (int i = 0; i < nb_read; i++) {
             buf[index++] = src[i];
             index = index % capacity;
         }
